proto_tech3-main/src/stripe_/requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
stripe==12.2.0
python-dotenv==1.1.1
email-validator==2.2.0

# Optional dependencies
sentry-sdk[fastapi]==2.32.0